Now the root and login bloc are wrapping the material app.
With this change now we can reach the blocs in the method "didChangeDependencies" in the stateful widgets.
"Items" option have been created to see the item and category list.
The item detail page design is done, to finish it, we need to can edit, create and delete (logicaly) an item.
The color palette is in the root bloc.