Item model have changes.
Attributes categoryId and measureId of item model have been deleted, instead they have been changed by "category" and "measure" each one respectively.
In this way it is easier to control item relationship, avoiding to call a function to know what category or measure is each one.
In the Item Detail page have been implemented dropdown options for its category and measure.
Services of "category" has been modified.
Services of "measure" has been created and added to the inventory repository.
When the item api is called to get its data, in the begin we are setting all the streams with a null value. To empty the form.