Item update is done.
Now we can change the state of the "Active" or "Inactive".
The form is ready to create a item but the bloc is incomplete.
We can call "create item page" from the list of items and the search page of the cash page.
When the parameter "item" is null on the Item's details widget, that means we are creating an item.
It have been implemented the dispose method to close the item's bloc.
The service to call a list of items by category just calling the item with state active ("A").