Item can by created from Cash Page with the category.
Now you can 'delete' the item. Actually it only change its state to 'E' (logical delete).
We need to hide items and any model that manage state attributes what its value is 'E'.
It start to develop the 'crud' of category, it needs to add an attribute to control its state.
Services like 'create' and 'update' for category have been created and the bloc is almost ready (category_bloc).