CRUD of category is done.
Category now have a attribute to set the state, active, inactive or deleted (logically).
From the method "toFireJson" we remove the attribute 'id', because it's a representation of documentID of firebase. If we keep it, we will be creating a new field on the base.
We create "CashCheckOut" page, this is the page to fill the user data, and make the check out.
Developing an animation with hero widget to move the invoice detail page from right to left when confirm the invoice detail, after pick up all the items to pay.
We remove the option to create a category from the cash page.