So close to finish the check out of the invoice.
Check out page is created, we can specify the cash with what we are paying the invoice.
We need to ask the user if he wants to add a customer to the invoice or not. With that we need to create an interface to search an user or create a new one.
We need to control the behavior when you want to back in the last part of the invoice, instead to press "new invoice" button.
Buttons to add to payment from credit card or check are just IU.
Customer model have been created and added to the invoice model.