Suggestion list for the customer in the invoice have changes.
Now it offer a button to create a customer from it (UI).
Customer invoice page have information about its shopping summary (UI).
A button from customer details let you edit customer information (UI).
We need to incorporate a function to show the last buyers of the shop when you are looking to add a customer to the invoice.