Change in the structure tree.
We have experiment some issues when you want to log out from other page that not be cash page.
The app is nested by root and authentication bloc correctly.
Renamed logon bloc to authentication bloc.
We need to fix a issue in the root page, when we are calling it to log out, the login page is called twice from the stream builder, apparently the stream is loading with 2 nulls, and the login page is been called for each one of them.