The new invoice call was fixed.
We had an issue, when an invoice is done if the user press the back button, the invoice's data persisted, now we validate in the dispose method of the "check_out" if the invoice is processed, if it's true, we call the "new invoice" method to set null all the streams.
The user drawer now is established on the "cash_page" scaffold and a global key of this, is sent to the others to be called from them, setting up in the top.