Enterprise model have been created.
The enterprise entity have been incorporated to the user model, in the authentication services we are getting the enterprise for the user.
We changed the background image and icon avatar for the user drawer.
When we press the "POS" option in the user drawer it not call the page, it just hide.
Now we can show the items in the cash page by list or by images.
When we "continue" in the process of the invoice because we want to check out, in the init state, we call the function to set the process status of the invoice to false.