Enterprise model have been incorporated to the other models.
Now when we call roles, items, categories and invoice, we are filtering by the enterprise.
All the models have a relationship with the enterprise. In the back end an att call "enterpriseId" is created.
A "transaction" model have been created with a static att to control the states of transactions, and a function to call it.