We create a "root page" in the POS.
The idea is that page could be a filter, if we have opened a cash drawer, then you go to the cash page, otherwise you need to open a cash drawer with this device.
We also need to create the user's business scheme, and company branches at the time of login, or link the device to the branch.