Changes in the login flow.
When you login the system will search if you have enterprise assigned.
If you only have one, it sends you directly to the home page. But if you have more then two, it let you pick you one.
After that it's gonna check for your role in this enterprise and finally go to the home page.
We need to incorporate if the device is register in some branch company, if not, we need to save the register with the relationship.
Finally if it's, it sends you to the home page, otherwise it'll ask you what is the default branch company for this one (it'll have to happen just once, the first time).