Validations for user and its enterprise are done.
When we login, the system check if the user have one enterprise or more assigned, if not, sending a error message.
The conditions for the enterprise of user are, state "A" and that the user has a role.
If the user have been assigned only to one enterprise, it's send directly to the "home page".
Authentication Root Page have been created to control the authentication flow via stream.
The message listener what was in the login page, have been relocated on the authentication root page for context issues.
The main file create all below a MaterialApp, we need to check this when we would go to work with routes.
In the device model we incorporate the branch office. We need to work on it.
Services of the device have been created.
The root page have been deleted, authentication root took its paper.