Authentication flow in the login is complete.
When we create a device, we specify the id that it will take.
The page to pick a branch if the device does not have one assigned is done.
Have been added to the branch model of the state attribute, to only fetch the "assets" branch offices.
In the device model we change the name of the attributes of auditory and the attribute type of the creation / modification date, from string to date.