Log out is done.
The home page is called by a material app to control the flow when we log out. It means that we have 2 material apps chained, one for the authentication page, and the last one when finally get in the app.
We add options to the user drawer for the Orders and Recipes.