Changes in the invoice model
We delete the enterprise field from the invoice, and replace that with the branch. This triggers a series of changes in services, repositories and others.
We create a home page for POS and Orders.
POS home page is just created.
Orders home page show a list of orders created to see its detail, and will let you edit and create new one (working on it).