Order Home Page is done, UI aspect could be better.
We need to work on the CRUD page (now called CashPage). It at the same time have to work with invoice, table orders and external orders.
"FromFireJson" function in the invoice model has been corrected.
The observable attribute branchList in the authentication pad, had to pass to ValueObservable because we need to use it on the orders home page for fill a dropdown menu.