We can create a invoice and a order from their programs.
Fixed issues filtering the orders by date, "to date" was the problem.
We add the attribute "note" to the invoice model.
We can enter in the cash page with order data to modify it, but the last process, upload the changes, we need to implemented it.
Cash page now receive the invoice parameter, to work as a "update mode".
"Total value" was not showing, the issue was on the "fromFireJson" method which was trying to take the value but with other name, "totalDetail" instead "total".
Order detail page have been created to show the details of the order, we need to keep working here to add more attributes and payments.
We need to change the background color of the date dialogs.