Update detail of the invoice, is done.
Now we can add discount to the invoice detail, we just need to tap on the item in the invoice detail and then it will show dialog show up to form, where we can change quantity and discounts on the detail.
Now we need to create a UI where we can cancel to invoice document. And also, we need to save the number of invoice with totals when closing of the cash drawer.