Cancel documents is done.
Order home page was renamed to invoice home page, it gonna work for both models.
We can cancel an invoice/order from its page.
When we try to cancel an Invoice it checks if we have a crash drawer open with this device, if the invoice was made from this cash drawer and day, before to change its state to canceled.