Working in respond times of queries.
We filter the customer by its id.
Due we are having problems to load all the items in the item list, we are just getting the 10 first of them (or less if they are few).
In the Search Page we improve the respond time when you are looking for items of a category.
But it's not enought, we need to work in a back end, preferably with the Firebase functions.